+++ billchop-be/billchop-be.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 16.0.30523.141
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BillChopBE", "billchop-be\BillChopBE.csproj", "{07953E38-5741-4F59-9777-AD3A85C91483}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BillChopBETests", "BillChopBETests\BillChopBETests.csproj", "{EC0A7936-2435-461B-87AD-DC3AF5E69D39}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectPortableTools", "ProjectPortableTools\ProjectPortableTools.csproj", "{97184AEA-85D7-4563-9B41-F9B9EA52E58F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{07953E38-5741-4F59-9777-AD3A85C91483}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07953E38-5741-4F59-9777-AD3A85C91483}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07953E38-5741-4F59-9777-AD3A85C91483}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC0A7936-2435-461B-87AD-DC3AF5E69D39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC0A7936-2435-461B-87AD-DC3AF5E69D39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC0A7936-2435-461B-87AD-DC3AF5E69D39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC0A7936-2435-461B-87AD-DC3AF5E69D39}.Release|Any CPU.Build.0 = Release|Any CPU
+		{97184AEA-85D7-4563-9B41-F9B9EA52E58F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{97184AEA-85D7-4563-9B41-F9B9EA52E58F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{97184AEA-85D7-4563-9B41-F9B9EA52E58F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{97184AEA-85D7-4563-9B41-F9B9EA52E58F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
