+++ billchop-be/billchop-be/BillChopBE.csproj
@@ -26,4 +26,7 @@
     <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="5.6.3" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.2" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\ProjectPortableTools\ProjectPortableTools.csproj" />
+  </ItemGroup>
 </Project>
